backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-jose[cryptography]
python-multipart
alembic
cors
psycopg2-binary
pydantic[email]
